backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
pymongo==4.8.0
redis==5.0.8
celery==5.4.0
python-dotenv==1.0.1
pydantic==2.8.2
marshmallow==3.21.3
argon2-cffi==23.1.0
gunicorn==22.0.0
sentry-sdk==2.12.0
pytest==8.3.2
black==24.8.0
flake8==7.1.1
